MAT-7054 Fixed test cases and refactored code
